api/pyproject.toml
[tool.poetry]
name = "revealer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^5.29.0"
fastapi = "^0.75.2"
boto3 = "^1.22.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
